add batch processing
