+++ functions/dotnet/PowertoolsWorkshop.csproj
@@ -13,16 +13,17 @@
     <PackageReference Include="Amazon.Lambda.DynamoDBEvents" Version="2.2.0" />
     <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.3.1" />
     <PackageReference Include="Amazon.Lambda.S3Events" Version="3.0.1" />
-    <PackageReference Include="AWS.Lambda.Powertools.Idempotency" Version="0.1.0-preview" />
-    <PackageReference Include="AWS.Lambda.Powertools.Logging" Version="1.2.0" />
-    <PackageReference Include="AWS.Lambda.Powertools.Metrics" Version="1.3.3" />
-    <PackageReference Include="AWS.Lambda.Powertools.Parameters" Version="1.0.0" />
-    <PackageReference Include="AWS.Lambda.Powertools.Tracing" Version="1.2.0" />
-    <PackageReference Include="AWSSDK.AppSync" Version="3.7.200.34" />
-    <PackageReference Include="AWSSDK.Core" Version="3.7.202.3" />
-    <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.7.201.5" />
-    <PackageReference Include="AWSSDK.Rekognition" Version="3.7.200.34" />
-    <PackageReference Include="AWSSDK.S3" Version="3.7.203.9" />
+    <PackageReference Include="AWS.Lambda.Powertools.BatchProcessing" Version="0.0.1-preview" />
+    <PackageReference Include="AWS.Lambda.Powertools.Idempotency" Version="0.2.2-preview" />
+    <PackageReference Include="AWS.Lambda.Powertools.Logging" Version="1.3.2" />
+    <PackageReference Include="AWS.Lambda.Powertools.Metrics" Version="1.4.2" />
+    <PackageReference Include="AWS.Lambda.Powertools.Parameters" Version="1.1.2" />
+    <PackageReference Include="AWS.Lambda.Powertools.Tracing" Version="1.3.2" />
+    <PackageReference Include="AWSSDK.AppSync" Version="3.7.200.42" />
+    <PackageReference Include="AWSSDK.Core" Version="3.7.202.11" />
+    <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.7.201.13" />
+    <PackageReference Include="AWSSDK.Rekognition" Version="3.7.200.42" />
+    <PackageReference Include="AWSSDK.S3" Version="3.7.203.17" />
     <PackageReference Include="AWSXRayRecorder.Handlers.AwsSdk" Version="2.12.0" />
     <PackageReference Include="SkiaSharp.NativeAssets.Linux.NoDependencies" Version="2.88.5" />
   </ItemGroup>
