+++ functions/dotnet/module1/AppSyncOperations.cs
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: MIT-0
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
