+++ functions/dotnet/module2/ApiOperations.cs
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: MIT-0
+
 using System;
 using System.Net.Http;
 using System.Net.Mime;
